# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CampaignLanguage": {
        "description": "The languages Clipah writes supporting copy in.\n\nIndonesian first, because that is who this product is for. The values are the ISO 639-1\ncodes the transcript already records, so a Transcript's language selects copy without a\ntranslation table in between.",
        "enum": [
          "id",
          "en"
        ],
        "title": "CampaignLanguage",
        "type": "string"
      },
      "CampaignModelMetadataResponse": {
        "additionalProperties": false,
        "description": "What produced this copy, so a member can tell how much to trust it.",
        "properties": {
          "deterministic": {
            "title": "Deterministic",
            "type": "boolean"
          },
          "generator": {
            "title": "Generator",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "generator",
          "version",
          "deterministic"
        ],
        "title": "CampaignModelMetadataResponse",
        "type": "object"
      },
      "CampaignOutputListResponse": {
        "additionalProperties": false,
        "description": "The copy derived from one Edit, ordered so two readings can be compared.",
        "properties": {
          "campaignOutputs": {
            "items": {
              "$ref": "#/components/schemas/CampaignOutputResponse"
            },
            "title": "Campaignoutputs",
            "type": "array"
          }
        },
        "required": [
          "campaignOutputs"
        ],
        "title": "CampaignOutputListResponse",
        "type": "object"
      },
      "CampaignOutputResponse": {
        "additionalProperties": false,
        "description": "One complete piece of supporting copy, for one destination in one language.",
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "cta": {
            "title": "Cta",
            "type": "string"
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "title": "Hashtags",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "modelMetadata": {
            "$ref": "#/components/schemas/CampaignModelMetadataResponse"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "postCopy": {
            "title": "Postcopy",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "thumbnailBrief": {
            "$ref": "#/components/schemas/ThumbnailBriefResponse"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/CampaignWarningResponse"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "revision",
          "platform",
          "language",
          "title",
          "postCopy",
          "cta",
          "hashtags",
          "thumbnailBrief",
          "warnings",
          "modelMetadata",
          "createdAt"
        ],
        "title": "CampaignOutputResponse",
        "type": "object"
      },
      "CampaignRequestBody": {
        "additionalProperties": false,
        "description": "The three choices a member makes: which cut, for where, and in what language.",
        "properties": {
          "languages": {
            "items": {
              "$ref": "#/components/schemas/CampaignLanguage"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Languages",
            "type": "array"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Platforms",
            "type": "array"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "platforms",
          "languages"
        ],
        "title": "CampaignRequestBody",
        "type": "object"
      },
      "CampaignWarningResponse": {
        "additionalProperties": false,
        "description": "One caveat a member reads beside the copy, never instead of it.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "detail"
        ],
        "title": "CampaignWarningResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Platform": {
        "description": "The short-form destinations Clipah packages for, without publishing to any of them.",
        "enum": [
          "tiktok",
          "instagram_reels",
          "youtube_shorts"
        ],
        "title": "Platform",
        "type": "string"
      },
...
      "ThumbnailBriefResponse": {
        "additionalProperties": false,
        "description": "What a thumbnail has to carry, as the person making it reads it.",
        "properties": {
          "avoid": {
            "items": {
              "type": "string"
            },
            "title": "Avoid",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "visualDirection": {
            "title": "Visualdirection",
            "type": "string"
          }
        },
        "required": [
          "text",
          "visualDirection",
          "avoid"
        ],
        "title": "ThumbnailBriefResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clipah API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/google/callback": {
      "get": {
        "description": "Redeem the provider callback and start a Clipah Session for the resolved User.",
        "operationId": "complete_google_login_api_v1_auth_google_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Google Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/edits/{edit_id}/campaign-outputs": {
      "get": {
        "description": "List the copy already derived from this Edit's Revisions.",
        "operationId": "list_collection_api_v1_edits__edit_id__campaign_outputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Revision",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOutputListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Collection",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "description": "Derive supporting copy from one exact immutable Revision of this Edit.",
        "operationId": "create_api_v1_edits__edit_id__campaign_outputs_post",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignOutputListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "campaigns"
        ]
      }
    },
...
    "/api/v1/me/sessions/{session_id}": {
      "delete": {
        "description": "Revoke one of this User's own Sessions, or report it as absent.",
        "operationId": "revoke_one_session_api_v1_me_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke One Session",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}